Avoid integer promotion in readPin()
Apparently the result of _BV() is promoted to a signed integer, and this
sometimes results in suboptimal code generation, which can be noticeable
if it happens in a heavily used code.  Adding an explicit cast to
uint8_t prevents the compiler from inserting sign extension code.

Testing the change on the `xd87` keyboard with avr-gcc 8.4.0 shows that
the matrix scan rate increased from 6694 to 7014 scans per second (less
than 5%, but still an improvement).  For the `id80` keyboard with the
current code (converted to COL2ROW) the scan rate changes from 5776 to
5932 scans per second (less than 3%).  For the older version of the
`id80` code with the ROW2COL matrix the scan rate changes from 3866 to
3989 scans per second (also about 3%).

+++ platforms/avr/gpio.h
@@ -33,6 +33,6 @@ typedef uint8_t pin_t;
 #define writePinLow(pin) (PORTx_ADDRESS(pin) &= ~_BV((pin)&0xF))
 #define writePin(pin, level) ((level) ? writePinHigh(pin) : writePinLow(pin))
 
-#define readPin(pin) ((bool)(PINx_ADDRESS(pin) & _BV((pin)&0xF)))
+#define readPin(pin) ((bool)(PINx_ADDRESS(pin) & (uint8_t)_BV((pin)&0xF)))
 
 #define togglePin(pin) (PORTx_ADDRESS(pin) ^= _BV((pin)&0xF))
